+++ src/shims/fancybox.js
@@ -1,11 +1,11 @@
-import { loader } from '../util';
+import { rule } from '../util';
 
 
-export default loader(
+export default rule(
     {
         test: /\/jquery.fancybox.js$/,
         include: /\/bower_components\/fancybox\/source\//,
-        loader: 'imports?' +
+        loader: 'imports-loader?' +
             'window=>global,document=>window.document,jQuery=jquery'
     }
 );